simple_app/requirements.txt
fastapi
uvicorn
jinja2
sqlalchemy
scikit-learn
python-multipart
pydantic
markdown
pillow
pypdf
pymupdf
mammoth
striprtf
bleach
mutagen
pandas
openpyxl
xlrd==1.2.0
httpx
argon2-cffi
python-jose
itsdangerous
email-validator
psycopg2-binary
